% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncomp.R
\name{pca_broken_stick}
\alias{pca_broken_stick}
\title{Broken-stick expected proportions}
\usage{
pca_broken_stick(p)
}
\arguments{
\item{p}{Number of components/dimensions.}
}
\value{
Expected variance proportions under the broken-stick model.
}
\description{
Broken-stick expected proportions
}
\examples{
bs <- pca_broken_stick(8)
bs
sum(bs)  # proportions sum to one
}
