% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structured.R
\name{pca_incremental_init}
\alias{pca_incremental_init}
\title{Initialize an incremental PCA state}
\usage{
pca_incremental_init(p, center = TRUE, scale = FALSE)
}
\arguments{
\item{p}{Number of variables.}

\item{center}{Center data in the final model.}

\item{scale}{Scale data to unit variance in the final model.}
}
\description{
Initialize an incremental PCA state
}
\examples{
# Streaming PCA over three chunks
set.seed(9)
X <- matrix(rnorm(600), 150, 4)
st <- pca_incremental_init(4, center = TRUE, scale = TRUE)
st <- pca_incremental_update(st, X[1:50, ])
st <- pca_incremental_update(st, X[51:100, ])
st <- pca_incremental_update(st, X[101:150, ])
fit <- pca_incremental_finalize(st, ncomp = 2)
fit$eigenvalues
}
