% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teaching.R
\name{pca_teach_biplot}
\alias{pca_teach_biplot}
\title{Teach biplot geometry}
\usage{
pca_teach_biplot(fit = NULL, dims = c(1, 2))
}
\arguments{
\item{fit}{Optional PCA fit.}

\item{dims}{Two components.}
}
\description{
Teach biplot geometry
}
\examples{
t <- pca_teach_biplot()
dim(t$data$scores)
dim(t$data$loadings)
}
