% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teaching.R
\name{pca_teach_eigen}
\alias{pca_teach_eigen}
\title{Teach eigenvalues and eigenvectors in PCA}
\usage{
pca_teach_eigen(data = NULL, scale = FALSE)
}
\arguments{
\item{data}{Numeric data.}

\item{scale}{Standardize before building the covariance/correlation matrix.}
}
\description{
Teach eigenvalues and eigenvectors in PCA
}
\examples{
t <- pca_teach_eigen(iris[, 1:4], scale = TRUE)
t$data$eigenvalues
max(t$data$eigen_equation_residual)
}
