% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teaching.R
\name{pca_teach_loadings}
\alias{pca_teach_loadings}
\title{Teach loadings and variable-component relationships}
\usage{
pca_teach_loadings(fit = NULL)
}
\arguments{
\item{fit}{Optional PCA fit.}
}
\description{
Teach loadings and variable-component relationships
}
\examples{
t <- pca_teach_loadings()
t$data$loadings[, 1:2]
t$data$cos2[, 1:2]
}
