% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teaching.R
\name{pca_teach_ncomp}
\alias{pca_teach_ncomp}
\title{Teach component-number selection}
\usage{
pca_teach_ncomp(data = NULL, nperm = 99L)
}
\arguments{
\item{data}{Optional numeric data.}

\item{nperm}{Number of reference datasets for the demonstration.}
}
\description{
Teach component-number selection
}
\examples{
t <- pca_teach_ncomp(nperm = 19)
t$data$recommendations
}
