% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teaching.R
\name{pca_teach_scaling}
\alias{pca_teach_scaling}
\title{Teach the effects of scaling}
\usage{
pca_teach_scaling(data = NULL)
}
\arguments{
\item{data}{Optional numeric data.}
}
\description{
Teach the effects of scaling
}
\examples{
t <- pca_teach_scaling()
t$data$covariance_pca$eigenvalues
t$data$correlation_pca$eigenvalues
}
