% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teaching.R
\name{pca_teach_scores}
\alias{pca_teach_scores}
\title{Teach scores}
\usage{
pca_teach_scores(fit = NULL)
}
\arguments{
\item{fit}{Optional PCA fit.}
}
\description{
Teach scores
}
\examples{
t <- pca_teach_scores()
head(t$data)
}
