% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teaching.R
\name{pca_teach_svd}
\alias{pca_teach_svd}
\title{Teach singular value decomposition and its equivalence to PCA}
\usage{
pca_teach_svd(data = NULL, scale = FALSE)
}
\arguments{
\item{data}{Numeric data.}

\item{scale}{Standardize variables.}
}
\description{
Teach singular value decomposition and its equivalence to PCA
}
\examples{
t <- pca_teach_svd(iris[, 1:4], scale = TRUE)
head(t$data$V)
t$data$eigenvalues
}
