% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teaching.R
\name{pca_teach_variance}
\alias{pca_teach_variance}
\title{Teach variance maximization over projection directions}
\usage{
pca_teach_variance(data = NULL, angles = seq(0, 180, by = 1))
}
\arguments{
\item{data}{Two-column numeric data.}

\item{angles}{Candidate angles in degrees.}
}
\description{
Teach variance maximization over projection directions
}
\examples{
t <- pca_teach_variance()
head(t$data)
t$data$angle[which.max(t$data$projected_variance)]
}
