% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tour.R
\name{pca_tour_compare}
\alias{pca_tour_compare}
\title{Compare PCA view with local and guided tour histories}
\usage{
pca_tour_compare(fit, dims = c(1, 2), max_bases = 50L, angle = pi/8)
}
\arguments{
\item{fit}{A \code{pca_fit} object.}

\item{dims}{PCA dimensions used as the reference projection.}

\item{max_bases}{Number of saved bases per tour.}

\item{angle}{Local-tour radius.}
}
\description{
Compare PCA view with local and guided tour histories
}
\examples{
if (requireNamespace("tourr", quietly = TRUE)) {
  fit <- pca_fit(iris[, 1:4], center = TRUE, scale = TRUE, ncomp = 4)
  tc <- pca_tour_compare(fit, max_bases = 10)
  length(tc$local)
  length(tc$grand)
}
}
